refactor:Context構造をやめる
