+++ frontend/src/routes/__root.tsx
@@ -3,7 +3,6 @@ import type { QueryClient } from "@tanstack/react-query";
 import { createRootRouteWithContext, Outlet } from "@tanstack/react-router";
 import { TanStackRouterDevtoolsPanel } from "@tanstack/react-router-devtools";
 import { UserProvider } from "@/context/UserContext";
-import { PasswordProvider } from "@/context/PasswordContext";
 import TanStackQueryDevtools from "../integrations/tanstack-query/devtools";
 
 interface MyRouterContext {
@@ -15,9 +14,7 @@ export const Route = createRootRouteWithContext<MyRouterContext>()({
     <>
       <div className="w-full min-h-svh flex flex-col">
         <UserProvider>
-          <PasswordProvider>
-            <Outlet />
-          </PasswordProvider>
+          <Outlet />
         </UserProvider>
       </div>
       <TanStackDevtools
